backend/requirements.txt

flask
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
flask-cors
flask-session
pytest
Flask-Testing
